Add support for implementing diagnostics for `agent diagnose` (#14394)
* Addition of check's get_diagnose() method to support agent diangose CLI comamnd

* Add minor comment

* Fix linting issues

* Fix lint issues

* First iteration of diagnose API for integrations

This introduces the first design idea, providing an internal object
for integrations to easily access diagnose-related functionality, such
as reporting diagnoses as well as defining their own explicit
diagnostics. This removes the need to subclass `get_diagnoses()` (the
agent-facing API) and provides a convenient namespace for all the
functionality, without requiring explicit imports.

* Remove `DIAGNOSIS_NOT_ENABLE` enum value

* Handle exceptions during explicit diagnostics

We want to avoid for a single failing explicit diagnostic to prevent
the agent from retrieving other valid diagnosis results.

* Pass all diagnose fields correctly

* Sanitize fields

* Return JSON instead of Python objects from `get_diagnoses()`

* Rename raw_error -> rawerror to use correct field name

---------

Co-authored-by: Alex Lopez <alex.lopez@datadoghq.com> 3a8375e89e32b869d927ec4416eabbf150384aa9